Add logo, smooth out turning, give weapons and items weights and make the game respect them, add preliminary endgame screen, make quads share vertices.
git-svn-id: http://www.nemebean.com/svn/coldest@212 e6982656-166e-4d30-bf09-c8d9f77b066c

+++ Item.h
@@ -15,7 +15,7 @@ class Item
       int Uses() const{return usesleft;}
       float CoolMult() const{return coolmultiplier;}
       float AmmoMult() const{return ammomultiplier;}
-      float Weight() const{return weight;}
+      int Weight() const{return weight;}
       string Name() const{return name;}
       string ModelFile() const{return modelfile;}
       
@@ -33,7 +33,7 @@ class Item
       int type;
       float coolmultiplier;
       float ammomultiplier;
-      float weight;
+      int weight;
       string name;
       string modelfile;
